Install omegajail for the runner
This change refreshes the omegaup-backend bundle, since it included a
woefully out-of-date set of binaries due to a mistake in the uploading
script (it was clobbering the artifacts of omegaup-runner.tar.bz2
instead of omegaup-backend.tar.xz). It also downloads and installs
the omegajail bundle. Additionally, it makes all the necessary
subscribe/notify relationships so that if the backend bundle is
installed, all the necessary services reload.

Fixes: https://github.com/omegaup/omegaup/issues/1393